Include role for nodejs

+++ scripts/getroles.sh
@@ -25,16 +25,31 @@ __file="${__dir}/$(basename "${BASH_SOURCE[0]}")"
 __base="$(basename ${__file} .sh)"
 __root="$(dirname "${__dir}")"
 
-ansible-galaxy install \
- --force \
- --roles-path "${__root}/roles/deploy/v1.3.0" \
- carlosbuenosvinos.ansistrano-deploy,1.3.0
-mv "${__root}/roles/deploy/v1.3.0/carlosbuenosvinos.ansistrano-deploy/"* "${__root}/roles/deploy/v1.3.0/"
-rmdir "${__root}/roles/deploy/v1.3.0/carlosbuenosvinos.ansistrano-deploy/"
+if [ ! -f "${__root}/roles/deploy/v1.3.0/README.md" ]; then
+  ansible-galaxy install \
+   --force \
+   --roles-path "${__root}/roles/deploy/v1.3.0" \
+   carlosbuenosvinos.ansistrano-deploy,1.3.0
+  mv "${__root}/roles/deploy/v1.3.0/carlosbuenosvinos.ansistrano-deploy/"* "${__root}/roles/deploy/v1.3.0/"
+  rmdir "${__root}/roles/deploy/v1.3.0/carlosbuenosvinos.ansistrano-deploy/"
+fi
 
-ansible-galaxy install \
- --force \
- --roles-path "${__root}/roles/rollback/v1.2.0" \
- carlosbuenosvinos.ansistrano-rollback,1.2.0
- mv "${__root}/roles/rollback/v1.2.0/carlosbuenosvinos.ansistrano-rollback/"* "${__root}/roles/rollback/v1.2.0/"
- rmdir "${__root}/roles/rollback/v1.2.0/carlosbuenosvinos.ansistrano-rollback/"
+if [ ! -f "${__root}/roles/rollback/v1.2.0/README.md" ]; then
+  ansible-galaxy install \
+   --force \
+   --roles-path "${__root}/roles/rollback/v1.2.0" \
+   carlosbuenosvinos.ansistrano-rollback,1.2.0
+   mv "${__root}/roles/rollback/v1.2.0/carlosbuenosvinos.ansistrano-rollback/"* "${__root}/roles/rollback/v1.2.0/"
+   rmdir "${__root}/roles/rollback/v1.2.0/carlosbuenosvinos.ansistrano-rollback/"
+fi
+
+if [ ! -f "${__root}/roles/nodejs/v2.1.1/README.md" ]; then
+  ansible-galaxy install \
+   --force \
+   --roles-path "${__root}/roles/nodejs/v2.1.1" \
+   geerlingguy.nodejs,2.1.1
+   mv "${__root}/roles/nodejs/v2.1.1/geerlingguy.nodejs/"* "${__root}/roles/nodejs/v2.1.1/"
+   rmdir "${__root}/roles/nodejs/v2.1.1/geerlingguy.nodejs/"
+fi
+
+geerlingguy.nodejs
